refactor: Product api 명세 통일

+++ src/main/java/gift/config/DataInitializer.java
@@ -53,7 +53,7 @@ public class DataInitializer {
                 Category randomCategory = categories.get(random.nextInt(categories.size()));
                 Product product = new Product(
                     "Product " + i,
-                    random.nextLong(100000) + 1000,
+                    random.nextInt(100000) + 1000,
                     "https://img1.kakaocdn.net/thumb/C320x320@2x.fwebp.q82/?fname=https%3A%2F%2Fst.kakaocdn.net%2Fproduct%2Fgift%2Fproduct%2F20240508101036_6c7f02cb957848a69a25018a664a3c89.jpg",
                     randomCategory
                 );
